fix: support react-native
- export typed event emitter separately from setMaxListeners
- update react-native overrides in module package.json files
- add react-native-webrtc implementation to WebRTC transport
