+++ packages/interface/src/index.ts
@@ -16,7 +16,7 @@
 
 import type { Connection, NewStreamOptions, Stream } from './connection/index.js'
 import type { ContentRouting } from './content-routing/index.js'
-import type { TypedEventTarget } from './events.js'
+import type { TypedEventTarget } from './event-target.js'
 import type { Metrics } from './metrics/index.js'
 import type { PeerId } from './peer-id/index.js'
 import type { PeerInfo } from './peer-info/index.js'
@@ -690,5 +690,6 @@ export * from './stream-muxer/index.js'
 export * from './topology/index.js'
 export * from './transport/index.js'
 export * from './errors.js'
+export * from './event-target.js'
 export * from './events.js'
 export * from './startable.js'
