--- packages/stream-multiplexer-mplex/src/alloc-unsafe-browser.ts
@@ -1,3 +1,0 @@
-export function allocUnsafe (size: number): Uint8Array {
-  return new Uint8Array(size)
-}
