--- packages/stream-multiplexer-mplex/src/alloc-unsafe.ts
@@ -1,3 +1,0 @@
-export function allocUnsafe (size: number): Buffer {
-  return Buffer.allocUnsafe(size)
-}
