+++ packages/transport-webrtc/src/webrtc/index.react-native.ts
@@ -1,0 +1,5 @@
+export {
+  RTCIceCandidate,
+  RTCPeerConnection,
+  RTCSessionDescription
+} from 'react-native-webrtc'
